Add Tablayout and ViewPager to MediaLibraryActivity
