+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
             android:exported="false"
             android:screenOrientation="portrait"/>
         <activity
-            android:name=".library.ui.LibraryActivity"
+            android:name=".library.ui.MediaLibraryActivity"
             android:exported="false"
             android:screenOrientation="portrait"/>
         <activity
