+++ app/src/main/java/com/example/playlistmaker/library/ui/MediaLibraryActivity.kt
@@ -1,0 +1,45 @@
+package com.example.playlistmaker.library.ui
+
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import com.example.playlistmaker.R
+import com.example.playlistmaker.databinding.ActivityMediaLibraryBinding
+import com.google.android.material.tabs.TabLayoutMediator
+
+class MediaLibraryActivity : AppCompatActivity() {
+
+    private val binding by lazy {
+        ActivityMediaLibraryBinding.inflate(layoutInflater)
+    }
+
+    private lateinit var tabLayoutMediator: TabLayoutMediator
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContentView(binding.root)
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }
+
+        binding.toolbar.setNavigationOnClickListener {
+            finish()
+        }
+
+        binding.viewPager.adapter = MediaLibraryViewPagerAdapter(supportFragmentManager, lifecycle)
+
+        tabLayoutMediator =
+            TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
+                when (position) {
+                    0 -> tab.text = this.getString(R.string.favourite_tracks_title)
+                    1 -> tab.text = this.getString(R.string.playlists_title)
+                }
+            }
+        tabLayoutMediator.attach()
+    }
+}
