+++ app/src/main/res/values/strings.xml
@@ -39,4 +39,6 @@
     <string name="media_library_is_empty">Ваша медиатека пуста</string>
     <string name="new_playlist">Новый плейлист</string>
     <string name="no_playlists">Вы не создали\nни одного плейлиста</string>
+    <string name="favourite_tracks_title">Избранные треки</string>
+    <string name="playlists_title">Плейлисты</string>
 </resources>
